cleanup bootstrap policy to allow future changes
